+++ mtg_proxy_printer/print.py
@@ -25,7 +25,7 @@ from pathlib import Path
 from threading import BoundedSemaphore
 import typing
 
-from PySide6.QtCore import QObject, QMarginsF, QSizeF, Slot, QPersistentModelIndex, QThreadPool
+from PySide6.QtCore import QObject, QMarginsF, QSizeF, QSize, Slot, QPersistentModelIndex, QThreadPool
 from PySide6.QtGui import QPainter, QPdfWriter, QPageSize, QImage, QColor
 from PySide6.QtPrintSupport import QPrinter
 
@@ -40,7 +40,7 @@ from mtg_proxy_printer.settings import settings
 from mtg_proxy_printer.model.document import Document
 from mtg_proxy_printer.model.page_layout import PageLayoutSettings
 from mtg_proxy_printer.model.carddb import with_database_write_lock
-from mtg_proxy_printer.ui.page_scene import RenderMode, PageScene
+from mtg_proxy_printer.page_scene.page_scene import RenderMode, PageScene
 from mtg_proxy_printer.logger import get_logger
 import mtg_proxy_printer.units_and_sizes
 logger = get_logger(__name__)
@@ -60,7 +60,7 @@ PNGEncoderThreadLimit = BoundedSemaphore(max(1, process_cpu_count()-1))
 
 
 class PNGRenderer(ProgressSignalContainer):
-    def __init__(self, main_window: "MainWindow", document: Document, file_path: str):
+    def __init__(self, main_window: "MainWindow|None", document: Document, file_path: str):
         super().__init__(main_window)
         self.document = document
         self.file_path = Path(file_path)
@@ -80,22 +80,31 @@ class PNGRenderer(ProgressSignalContainer):
             round(RESOLUTION.magnitude))
         pool = QThreadPool.globalInstance()
         scene = PageScene(document, RenderMode.ON_PAPER, self)
+        dots_per_meter = round(RESOLUTION.to("pixel/meter").magnitude)
+        background_color = settings["export"].get_color("png-background-color")
         number_width = len(str(page_count))
         parent = file_path.parent
         self.begin_update.emit(page_count, self.tr("Export as PNGs"))
         for page_nr in range(page_count):
             file_name = f"{file_path.stem}-{str(page_nr + 1).zfill(number_width)}.png"
             output_path = str(parent / file_name)
-            background_color = QColor(settings["export"]["png-background-color"])
-            # 255 is solid. So avoid adding the alpha channel, if it won't be used.
-            image_format = Format.Format_RGB888 if background_color.alpha() == 255 else Format.Format_RGBA8888
-            image = QImage(page_size, image_format)
-            image.fill(background_color)
+            image = self._create_image(page_size, background_color, dots_per_meter)
             painter = QPainter(image)
             page_index = QPersistentModelIndex(document.index(page_nr, 0))
             scene.on_current_page_changed(page_index)
             scene.render(painter)
+            painter.end()
             pool.start(partial(self._compress_single_image, image, output_path))
+
+    @staticmethod
+    def _create_image(page_size: QSize, background_color: QColor, dots_per_meter: int):
+        # 255 is solid. So avoid adding the alpha channel, if it won't be used.
+        image_format = Format.Format_RGB888 if background_color.alpha() == 255 else Format.Format_RGBA8888
+        image = QImage(page_size, image_format)
+        image.setDotsPerMeterX(dots_per_meter)
+        image.setDotsPerMeterY(dots_per_meter)
+        image.fill(background_color)
+        return image
 
     @with_database_write_lock(PNGEncoderThreadLimit)
     def _compress_single_image(self, image: QImage, output_path: str):