+++ mtg_proxy_printer/settings.py
@@ -176,6 +176,9 @@ VALID_CUT_MARKER_STYLES: defaultdict[str, PenStyle] = defaultdict(lambda: PenSty
     "Dots": PenStyle.DotLine,
     "Dashes": PenStyle.DashLine,
 })
+VALID_PRINT_REGISTRATION_MARKS_STYLES: set[str] = {
+    "None", "Bullseye",
+}
 DEFAULT_MARGINS = 5*mm
 DEFAULT_SETTINGS["documents"] = {
     "card-bleed": "0 mm",
@@ -191,6 +194,7 @@ DEFAULT_SETTINGS["documents"] = {
     "margin-bottom": str(DEFAULT_MARGINS),
     "margin-left": str(DEFAULT_MARGINS),
     "margin-right": str(DEFAULT_MARGINS),
+    "print-registration-marks-style": "None",
     "row-spacing": "0 mm",
     "column-spacing": "0 mm",
     "print-sharp-corners": "False",
@@ -393,7 +397,9 @@ def _validate_documents_section(to_validate: ConfigParser, section_name: str = "
         section["default-document-name"] = document_name[:MAX_DOCUMENT_NAME_LENGTH-1] + "…"
     defaults = DEFAULT_SETTINGS[section_name]
     boolean_settings = {"print-sharp-corners", "print-page-numbers", "cut-marker-draw-above-cards",}
-    string_settings = {"default-document-name", "paper-size", "paper-orientation", "watermark-text", "cut-marker-style",}
+    string_settings = {
+        "default-document-name", "paper-size", "paper-orientation", "watermark-text",
+        "cut-marker-style", "print-registration-marks-style"}
     color_settings = {"watermark-color", "cut-marker-color",}
     for key in section.keys():
         if key in DOCUMENT_SETTINGS_QUANTITY_LIMITS:
@@ -413,6 +419,8 @@ def _validate_documents_section(to_validate: ConfigParser, section_name: str = "
         _restore_default(section, defaults, "paper-size")
     if section["paper-orientation"] not in PageSizeManager.PageOrientation:
         _restore_default(section, defaults, "paper-orientation")
+    if section["print-registration-marks-style"] not in VALID_PRINT_REGISTRATION_MARKS_STYLES:
+        _restore_default(section, defaults, "print-registration-marks-style")
     # Check some semantic properties
     available_height = section.get_quantity("custom-page-height") - \
         (section.get_quantity("margin-top") + section.get_quantity("margin-bottom"))